# artifacts/handoff-w4-round-1/provider-artifacts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/organizations/{organizationId}/programs": {
      "get": {
        "operationId": "ProgramsController_list",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      },
      "post": {
        "operationId": "ProgramsController_create",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/templates": {
      "get": {
        "operationId": "ProgramsController_templates",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}": {
      "get": {
        "operationId": "ProgramsController_get",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      },
      "patch": {
        "operationId": "ProgramsController_update",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}/versions": {
      "get": {
        "operationId": "ProgramsController_versions",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}/versions/{versionId}": {
      "get": {
        "operationId": "ProgramsController_version",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "versionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}/draft": {
      "post": {
        "operationId": "ProgramsController_draft",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}/draft/abandon": {
      "post": {
        "operationId": "ProgramsController_abandonDraft",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}/preview": {
      "get": {
        "operationId": "ProgramsController_preview",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "layout",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}/validate": {
      "post": {
        "operationId": "ProgramsController_validate",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}/test-sessions": {
      "post": {
        "operationId": "ProgramsController_createTestSession",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/test-sessions/{sessionId}": {
      "get": {
        "operationId": "ProgramsController_getTestSession",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/test-sessions/{sessionId}/stamps": {
      "post": {
        "operationId": "ProgramsController_addStamps",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/test-sessions/{sessionId}/redeem/{rewardId}": {
      "post": {
        "operationId": "ProgramsController_redeem",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "rewardId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/test-sessions/{sessionId}/reverse": {
      "post": {
        "operationId": "ProgramsController_reverse",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/test-sessions/{sessionId}/reset": {
      "post": {
        "operationId": "ProgramsController_reset",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sessionId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}/publish": {
      "post": {
        "operationId": "ProgramsController_publish",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}/pause": {
      "post": {
        "operationId": "ProgramsController_pause",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}/resume": {
      "post": {
        "operationId": "ProgramsController_resume",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}/archive": {
      "post": {
        "operationId": "ProgramsController_archive",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
    "/v1/organizations/{organizationId}/programs/{programId}/restore": {
      "post": {
        "operationId": "ProgramsController_restore",
        "parameters": [
          {
            "name": "organizationId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "programId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Programs"
        ]
      }
    },
...
    "/v1/public/programs/{programSlug}": {
      "get": {
        "operationId": "PublicEnrollmentController_program",
        "parameters": [
          {
            "name": "programSlug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tenant",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "PublicEnrollment"
        ]
      }
    },
...
  },
  "info": {
    "title": "Waflo Platform API",
    "description": "Authoritative Waflo W4 API for merchant administration and signed Staff devices.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "waflo_session"
      }
    },
...
  }
}